sound: drop dead/gratuitous FP from the int16 audio output path
The libretro audio path is int16 end-to-end (int32 accumulators in the
Wave/WaveHi buffers, packed L|R int16 to the frontend), so any floating
point in a function that returns audio samples is pure waste. Two spots
qualified:

emu2413 (vendored YM2413/OPLL): OPLL_calc carried a double sub-sample
interpolator gated on opll->quality. fceumm calloc-zeroes quality and
never calls OPLL_set_quality, so both VRC7 mix paths already resolved to
the integer calc() - the interpolator was unreachable and the only FP in
any audio-returning function. Removed; OPLL_calc now just returns calc().
The realstep/opllstep/oplltime/prev/next/out fields are kept because the
VRC7 mapper serialises them; they stay zero as before, so the savestate
layout is byte-identical and output is bit-identical.

n106 (Namco 163): the EG-cache register write computed (double)(V&0xF)*
576716. (V&0xF) is 0..15 and the product (<=8650740) fits int32 exactly,
so the double was decorative - replaced with the integer expression.

Per-sample output in both TUs is now FP-free (verified via objdump: no
cvtsi2sd/mulsd/etc. in calc/OPLL_calc/OPLL_fillbuf or the N163 register
write). Remaining doubles are one-shot init table/rate generation
(maketables, the per-fill N163 inc), which are deterministic by design
and not a float->int16 round trip.

+++ src/boards/n106.c
@@ -157,7 +157,10 @@ static void FixCache(int a, int V) {
 		/* fix be like in https://github.com/SourMesen/Mesen/blob/cda0a0bdcb5525480784f4b8c71de6fc7273b570/Core/Namco163Audio.h#L61 */
 		LengthCache[w] = 256 - (V & 0xFC);
 		break;
-	case 0x07: EnvCache[w] = (double)(V & 0xF) * 576716; break;
+	/* (V & 0xF) is 0..15; 15 * 576716 = 8650740 fits an int32 exactly, so
+	 * the old (double) multiply produced an integer result anyway - compute
+	 * it directly to keep this audio TU free of floating point. */
+	case 0x07: EnvCache[w] = (uint32_t)(V & 0xF) * 576716; break;
 	}
 }
